fix x64 build with VS2015 UP2 compiler (#640)
fix VS2015 compiler warnings for: address-model=64, deprecated-functions=on, link=shared

+++ examples/print.cpp
@@ -464,7 +464,7 @@ void print(char const* buf)
 	{
 		if (*buf == '\033' && buf[1] == '[')
 		{
-			WriteFile(out, start, buf - start, &written, NULL);
+			WriteFile(out, start, DWORD(buf - start), &written, NULL);
 			buf += 2; // skip escape and '['
 			start = buf;
 			if (*buf == 0) break;
@@ -533,7 +533,7 @@ one_more:
 			++buf;
 		}
 	}
-	WriteFile(out, start, buf - start, &written, NULL);
+	WriteFile(out, start, DWORD(buf - start), &written, NULL);
 
 #else
 	fputs(buf, stdout);
